Resolve LinkTo circular dependencies
