+++ client/src/components/LinkToNotPreviewed.js
@@ -122,8 +122,7 @@ export default class LinkTo extends Component {
     } else {
       to = model
     }
-    console.dir("componentProps")
-    console.dir(componentProps)
+
     const LinkToComponent = componentClass
     return (
       <LinkToComponent to={to} style={style} {...componentProps}>
