+++ client/src/pages/people/Show.js
@@ -5,12 +5,12 @@ import AppContext from "components/AppContext"
 import AssessmentResultsContainer from "components/assessments/AssessmentResultsContainer"
 import AssignPositionModal from "components/AssignPositionModal"
 import AvatarDisplayComponent from "components/AvatarDisplayComponent"
-import { mapReadonlyCustomFieldsToComps } from "components/CustomFields"
+import { mapReadonlyCustomFieldsToComps } from "components/CustomFieldsReadonly"
 import EditAssociatedPositionsModal from "components/EditAssociatedPositionsModal"
+import { parseHtmlWithLinkTo } from "components/editor/LinkAnet"
 import * as FieldHelper from "components/FieldHelper"
 import Fieldset from "components/Fieldset"
 import GuidedTour from "components/GuidedTour"
-import { parseHtmlWithLinkTo } from "components/editor/LinkAnet"
 import LinkTo from "components/LinkTo"
 import Messages from "components/Messages"
 import { DEFAULT_CUSTOM_FIELDS_PARENT } from "components/Model"
@@ -428,7 +428,7 @@ const PersonShow = ({ pageDispatchers }) => {
                 Settings.dateFormats.forms.displayShort.date
               ),
             role: Person.humanNameOfRole(values.role),
-            biography: parseHtmlWithLinkTo(person.biography)
+            biography: parseHtmlWithLinkTo(person.biography, LinkTo)
           }
           return Person.shownStandardFields.reduce((accum, key) => {
             accum[key] = (
